backend/requirements.txt
Django==2.2.12
psycopg2==2.8.5
django-storages==1.9.1
djangorestframework==3.11.0